backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.0
lxml==4.9.3
httpx==0.25.2
playwright==1.40.0
